backend/requirements.txt
flask
flask-cors
python-dotenv
openai
langchain
chromadb
torch
torchvision
transformers
accelerate
huggingface-hub==0.23.2
librosa
pydub
soundfile
requests
numpy
pandas
deep-translator
nemo_toolkit
edge-tts
pyaudio
torchaudio
timm
pillow
einops
peft
sentencepiece
bitsandbytes
protobuf
decord
git+https://github.com/huggingface/parler-tts.git
indictranstoolkit==1.1.1
hydra-core